@@ -24,32 +24,35 @@
   "babel": {
     "env": {
       "production": {
-        "presets": ["es2015-rollup"]
+        "presets": [
+          "@babel/preset-env"
+        ]
       },
       "development": {
-        "presets": ["es2015"]
+        "presets": [
+          "@babel/preset-env"
+        ]
       }
     }
   },
   "license": "MIT",
   "dependencies": {
-    "rxjs": "^6.1.0",
-    "rxjs-compat": "^6.1.0"
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-es2015-rollup": "^3.0.0",
-    "babelrc-rollup": "^3.0.0",
+    "@babel/core": "^7.1.2",
+    "@babel/node": "^7.0.0",
+    "@babel/preset-env": "^7.1.0",
+    "babel-preset-env": "^1.7.0",
     "esdoc": "^1.1.0",
     "esdoc-standard-plugin": "^1.0.0",
     "istanbul": "^1.1.0-alpha.1",
-    "rollup": "^0.58.2",
-    "rollup-plugin-babel": "^3.0.2",
-    "tap-dot": "^1.0.5",
-    "tape": "^4.8.0",
-    "typescript": "^2.5.3"
+    "rollup": "^0.66.2",
+    "rollup-plugin-babel": "^4.0.3",
+    "tap-dot": "^2.0.0",
+    "tape": "^4.9.1",
+    "typescript": "^3.1.1"
   },
   "types": "./ts/rxmq.d.ts"
 }
